+++ src/app/services/course.service.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
-import { Observable, catchError, tap } from 'rxjs';
+import { Observable, catchError, tap, throwError } from 'rxjs';
 import { Course } from '../models/Course';
 import { Response } from '../models/Response';
 import { environment } from '../../environments/environment';
+import { UserService } from './user.service';
 
 @Injectable({
   providedIn: 'root'
@@ -12,10 +13,12 @@ export class CourseService {
   private baseApiUrl = environment.baseApiUrl;
   private apiUrl = `${this.baseApiUrl}/api/curso`;
 
-  constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient, private userService: UserService) { }
 
+  isLogged():boolean{
+    return this.userService.logged();
+  }
   getCourses(page: number, limit:number):Observable<Response<Course[]>>{
-
     // /posts?_page=7&_limit=10
     let params = new HttpParams()
       .set("_page", page)
@@ -28,14 +31,24 @@ export class CourseService {
   }
 
   createCourse(course: Course):Observable<Course>{
-    return this.http.post<Course>(this.baseApiUrl, course);
+    if(this.isLogged()){
+      return this.http.post<Course>(this.baseApiUrl, course);
+    }
+    return throwError(()=> new Error('Erro de autenticação'))
   }
 
   updateCourse(course: Course):Observable<Response<Course>> {
-    return this.http.put<Response<Course>>(`${this.baseApiUrl}/${course.id}`, course);
+    if(this.isLogged()){
+      return this.http.put<Response<Course>>(`${this.baseApiUrl}/${course.id}`, course);
+    }
+    return throwError(()=> new Error('Erro de autenticação'))
+    
   }
 
   deleteCourse(id:number):Observable<Course>{
-    return this.http.delete<Course>(`${this.baseApiUrl}/${id}`)
+    if(this.isLogged()){
+      return this.http.delete<Course>(`${this.baseApiUrl}/${id}`)
+    }
+    return throwError(()=> new Error('Erro de autenticação'))
   }
 }
